Completed Select Task
Refactored VisitManager to make use of two custom listeners.
ActiveVisitListener provides the prior functionality of VisitManager's
change method, and ActiveTaskListener provides similar functionality for
the active task. These classes are left package-private and not meant to
be a dependency for anything other than VisitManager.

Also, fixed a slight bug where the VisitView would try to open the
TaskView even if a Task had not been selected.